moving rating into component
